requirements.txt
{\rtf1\ansi\ansicpg1252\cocoartf2822
\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica;}
{\colortbl;\red255\green255\blue255;}
{\*\expandedcolortbl;;}
\paperw11900\paperh16840\margl1440\margr1440\vieww28600\viewh18000\viewkind0
\pard\tx720\tx1440\tx2160\tx2880\tx3600\tx4320\tx5040\tx5760\tx6480\tx7200\tx7920\tx8640\pardirnatural\partightenfactor0

\f0\fs24 \cf0 fastapi==0.104.1\
uvicorn==0.24.0\
sentence-transformers==2.2.2\
faiss-cpu==1.7.4\
pandas==2.0.3\
numpy==1.24.3\
requests==2.31.0\
beautifulsoup4==4.12.2\
pydantic==2.5.0}